+++ core/src/main/java/pl/agh/edu/ui/component/button/ScenarioButton.java
@@ -75,7 +75,7 @@ public class ScenarioButton extends WrapperContainer<Button> {
 	}
 
 	private void addSoundEvents() {
-		addListener(new InputListener() {
+		scenarioButton.addListener(new InputListener() {
 			@Override
 			public boolean touchDown(InputEvent event, float x, float y, int pointer, int button) {
 				if (!scenarioButton.isChecked()) {
